pyproject.toml
[project]
name = "portfolio-analysis-ai"
version = "0.1.0"
description = "Interactive portfolio analysis with AI-powered natural language queries"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "numpy==1.26.4",
    "pandas==1.5.3",
    "openai==1.59.8",
    "streamlit==1.41.1",
    "pandasai==1.4.10",
    "plotly==5.24.1",
    "matplotlib==3.10.0",
    "seaborn==0.13.2",
    "duckdb==0.8.1",
    "python-dotenv==1.0.1",
    "pydantic==1.10.8"
]

[project.optional-dependencies]
dev = [
    "black==24.1.0",
    "flake8==7.0.0",
    "pytest==8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 88
target-version = "py310"
